Make Sentry DSN configurable by file
PuppetServer (or at least, earlier versions of it) scrub the environment
variables before launching the Puppet Master under JRuby. This means
that `PUPPET_SENTRY_DSN` doesn't make it through to the Puppet runtime.
Later versions of Puppet Server let you get around this by specifying
the environment in one of the config files, but as we're on a legacy
version, we need another workaround.

This code looks for `sentry.conf` in the Puppet confdir (usually
`/etc/puppet`) and parses out key=value pairs. At the moment only the
key "dsn" is processed.